requirements.txt
flask
python-dotenv
requests
authlib
six
flask-cors